# packages/api-client/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/projects/{projectId}/environments": {
			"post": {
				"operationId": "projectEnvironmentsCreate",
				"parameters": [
					{
						"name": "projectId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateEnvironmentDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EnvironmentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Create an environment in a project",
				"tags": ["environments"]
			}
		},
		"/api/environments/{environmentId}": {
			"patch": {
				"operationId": "environmentsUpdate",
				"parameters": [
					{
						"name": "environmentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateEnvironmentDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EnvironmentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Update an environment",
				"tags": ["environments"]
			},
			"delete": {
				"operationId": "environmentsDelete",
				"parameters": [
					{
						"name": "environmentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EnvironmentOkDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Delete an environment",
				"tags": ["environments"]
			}
		},
...
		"/api/services/{serviceId}/deployments": {
			"get": {
				"operationId": "serviceDeploymentsList",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DeploymentViewDto"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List deployment history for a service",
				"tags": ["deployments"]
			},
			"post": {
				"operationId": "serviceDeploymentsEnqueue",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Queue a deployment for a service",
				"tags": ["deployments"]
			}
		},
		"/api/deployments/{deploymentId}": {
			"get": {
				"operationId": "deploymentsGet",
				"parameters": [
					{
						"name": "deploymentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get deployment details",
				"tags": ["deployments"]
			}
		},
		"/api/deployments/{deploymentId}/logs": {
			"get": {
				"operationId": "deploymentLogsList",
				"parameters": [
					{
						"name": "deploymentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentEventLogsDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List deployment event logs",
				"tags": ["deployments"]
			}
		},
		"/api/deployments/{deploymentId}/build-logs": {
			"get": {
				"operationId": "deploymentBuildLogsGet",
				"parameters": [
					{
						"name": "deploymentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentBuildLogsDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get deployment build logs",
				"tags": ["deployments"]
			}
		},
		"/api/deployments/{deploymentId}/cancel": {
			"post": {
				"operationId": "deploymentsCancel",
				"parameters": [
					{
						"name": "deploymentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Cancel a deployment",
				"tags": ["deployments"]
			}
		},
		"/api/invitations": {
			"get": {
				"operationId": "invitationsList",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/InvitationDto"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List invitations",
				"tags": ["invitations"]
			},
			"post": {
				"operationId": "invitationsCreate",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateInviteDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateInviteResultDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Create an invitation",
				"tags": ["invitations"]
			}
		},
		"/api/invitations/{id}": {
			"delete": {
				"operationId": "invitationsDelete",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InviteOkDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Revoke an invitation",
				"tags": ["invitations"]
			}
		},
		"/api/invitations/{id}/resend": {
			"post": {
				"operationId": "invitationsResend",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateInviteResultDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Resend an invitation",
				"tags": ["invitations"]
			}
		},
		"/api/invitations/{id}/validity": {
			"get": {
				"operationId": "invitationsValidity",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InviteValidityDto"
								}
							}
						}
					}
				},
				"summary": "Check invitation token validity",
				"tags": ["invitations"]
			}
		},
		"/api/invitations/{id}/accept": {
			"post": {
				"operationId": "invitationsAccept",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptInviteDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AcceptInviteResponseDto"
								}
							}
						}
					}
				},
				"summary": "Accept an invitation and log in",
				"tags": ["invitations"]
			}
		},
...
		"/api/platform/version/check": {
			"post": {
				"operationId": "platformVersionCheck",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlatformVersionCheckResultDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Check for platform updates",
				"tags": ["platform"]
			}
		},
...
	},
	"info": {
		"title": "kubwave API",
		"description": "",
		"version": "dev",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
...
			"SessionUserDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": ["id", "name", "email", "isAdmin"]
			},
...
			"EnvironmentDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"projectId": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"kind": {
						"type": "string",
						"enum": ["persistent", "preview"]
					},
					"prPreviewsEnabled": {
						"type": "boolean"
					},
					"prNumber": {
						"type": "number",
						"nullable": true
					},
					"prRepoUrl": {
						"type": "string",
						"nullable": true
					},
					"baseEnvironmentId": {
						"type": "string",
						"format": "uuid",
						"nullable": true
					},
					"serviceCount": {
						"type": "number"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"projectId",
					"name",
					"kind",
					"prPreviewsEnabled",
					"prNumber",
					"prRepoUrl",
					"baseEnvironmentId",
					"serviceCount",
					"createdAt",
					"updatedAt"
				]
			},
...
			"CreateEnvironmentDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					}
				},
				"required": ["name"]
			},
			"UpdateEnvironmentDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					},
					"prPreviewsEnabled": {
						"type": "boolean"
					}
				}
			},
			"EnvironmentOkDto": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": ["ok"]
			},
...
			"DeploymentViewDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"serviceId": {
						"type": "string",
						"format": "uuid"
					},
					"type": {
						"type": "string",
						"enum": ["docker-image", "dockerfile", "public-repo", "private-repo", "postgres", "mysql", "mariadb", "mongodb"]
					},
					"status": {
						"type": "string",
						"enum": ["pending", "deploying", "canceling", "succeeded", "failed", "superseded", "canceled"]
					},
					"phase": {
						"type": "string",
						"nullable": true
					},
					"lastError": {
						"type": "string",
						"nullable": true
					},
					"attempts": {
						"type": "number"
					},
					"config": {
						"type": "object",
						"additionalProperties": true
					},
					"trigger": {
						"type": "string",
						"enum": ["manual", "auto", "preview"]
					},
					"triggeredByUserId": {
						"type": "string",
						"format": "uuid",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"startedAt": {
						"type": "string",
						"nullable": true
					},
					"finishedAt": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"serviceId",
					"type",
					"status",
					"phase",
					"lastError",
					"attempts",
					"config",
					"trigger",
					"triggeredByUserId",
					"createdAt",
					"startedAt",
					"finishedAt"
				]
			},
			"DeploymentLogEntryDto": {
				"type": "object",
				"properties": {
					"ts": {
						"type": "string"
					},
					"level": {
						"type": "string",
						"enum": ["info", "warn", "error"]
					},
					"step": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": ["ts", "level", "step", "message"]
			},
			"DeploymentEventLogsDto": {
				"type": "object",
				"properties": {
					"logs": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/DeploymentLogEntryDto"
						}
					}
				},
				"required": ["logs"]
			},
			"DeploymentBuildLogContainerDto": {
				"type": "object",
				"properties": {
					"containerName": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"nullable": true
					}
				},
				"required": ["containerName", "content", "updatedAt"]
			},
			"DeploymentBuildLogsDto": {
				"type": "object",
				"properties": {
					"containers": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/DeploymentBuildLogContainerDto"
						}
					}
				},
				"required": ["containers"]
			},
			"InvitationDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"isAdmin": {
						"type": "boolean"
					},
					"invitedBy": {
						"type": "string",
						"format": "uuid",
						"nullable": true
					},
					"expiresAt": {
						"type": "string"
					},
					"acceptedAt": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": ["pending", "accepted", "expired"]
					}
				},
				"required": ["id", "email", "isAdmin", "invitedBy", "expiresAt", "acceptedAt", "createdAt", "status"]
			},
			"CreateInviteDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"isAdmin": {
						"type": "boolean",
						"default": false
					}
				},
				"required": ["email", "isAdmin"]
			},
			"CreateInviteResultDto": {
				"type": "object",
				"properties": {
					"invitation": {
						"$ref": "#/components/schemas/InvitationDto"
					},
					"emailSent": {
						"type": "boolean"
					},
					"emailError": {
						"type": "string"
					}
				},
				"required": ["invitation", "emailSent"]
			},
			"InviteOkDto": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": ["ok"]
			},
			"InviteValidityDto": {
				"type": "object",
				"properties": {
					"valid": {
						"type": "boolean"
					},
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": ["valid"]
			},
			"AcceptInviteDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1
					},
					"password": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": ["name", "password"]
			},
			"AcceptInviteResponseDto": {
				"type": "object",
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/SessionUserDto"
					}
				},
				"required": ["accessToken", "user"]
			},
...
			"PlatformVersionCheckResultDto": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": ["success", "message"]
			},
...
		}
	}
}